+++ finaev.alexey/P4/main.cpp
@@ -1,6 +1,7 @@
 #include <iostream>
 #include <stdexcept>
 #include <cstddef>
+#include <create_str.h>
 #include "string_conversion.h"
 
 int main()
