+++ finaev.alexey/P4/string_conversion.h
@@ -5,8 +5,6 @@
 
 namespace finaev
 {
-  void transferStr(char* new_str, const char* str, size_t length);
-  char* fillStr(std::istream& in);
   int checkRepeat(const char* str);
 }
